Aded example tests for blovking and async I2C slave.
The blocking tests are working fine for the ESP32C6 devices.

Issues remains with the async capability as it is difficult to manage clock stretching
in an async context. A workaround has been implemented for now, but further investigation
is needed to resolve the underlying driver issue.

An avenue that I'm looking at is to have an interrupt based approach for the I2C slave
driver, which would allow for better handling of clock stretching and async operations through
interrupts and callbacks handlers.